# Summary
Rewrite refdiff commit graph algorithm
<!--
Thanks for submitting a pull request!

We appreciate you spending the time to work on these changes.
Please fill out as many sections below as possible.
-->

### Key Points

- [ ] This is a breaking change
- [ ] New or existing documentation is updated

### Description
changed newRef and oldRef add place to fix the losting of them.
remove cayleygraph
add CommitNodeGraph
batch insert commits diffs by sql not for each one insert once.

graph perf cayleygraph and CommitNodeGraph.  one insert once and batch insert.
[new.txt](https://github.com/merico-dev/lake/files/8402413/new.txt)
[new_batch.txt](https://github.com/merico-dev/lake/files/8402414/new_batch.txt)
[old.txt](https://github.com/merico-dev/lake/files/8402415/old.txt)
[old_batch.txt](https://github.com/merico-dev/lake/files/8402416/old_batch.txt)

[In Graph Building]
![image](https://user-images.githubusercontent.com/2921251/161382058-19c64b62-9984-47ab-b49f-c2ef72fab2f8.png)
cayleygraph: 500ms-900ms
![image](https://user-images.githubusercontent.com/2921251/161382117-8bd097d3-36fb-47dd-b98b-9eed39657e47.png)
CommitNodeGraph 200ms-500ms
Our new algorithm quick a liitle because we don’t need create side length and other flags not used.

[In Calculate]
![image](https://user-images.githubusercontent.com/2921251/161382264-d99f6ab0-009f-4350-99e2-5fef8cb7a81e.png)
cayleygraph: 100ms - 200ms
![image](https://user-images.githubusercontent.com/2921251/161382324-921ae6f8-4712-4792-bd53-e04b15a4cf9b.png)
CommitNodeGraph : 100ms - 200ms
Our new algorithm same as the old.

[In Data Get]
![image](https://user-images.githubusercontent.com/2921251/161382410-c03bf403-3a23-4f32-ad87-de0754f14b31.png)
cayleygraph: 30000ms (30s)
![image](https://user-images.githubusercontent.com/2921251/161382455-78753d52-9665-48fb-b29e-725dd2e02d4e.png)
CommitNodeGraph: 1ms - 20ms

The iterator of the cayleygraph algorithm has a fatal flaw, and its time complexity is O(n^2) .
Our new commit node graph algorithm only time complexity O(n). So we are N times faster than before.(n is commits count)
[In DataBase Insert]
![image](https://user-images.githubusercontent.com/2921251/161382767-26642453-914b-48dc-92b2-34441ad69922.png)
Insert one by one:  10000ms+
![image](https://user-images.githubusercontent.com/2921251/161382834-2dba9b87-a0e7-4ea8-a5ca-c0dedc1b2e15.png)
Batch insert: 400ms - 1200ms

Less SQL request use less time.

[Ans]
![image](https://user-images.githubusercontent.com/2921251/161382936-706acd45-8dc5-4a36-b3d6-174080b065c6.png)
cayleygraph:6415
![image](https://user-images.githubusercontent.com/2921251/161382982-384a4bcb-7062-4df7-8419-db2c090a7e35.png)
CommitNodeGraph:6415

The answer between cayleygraph and CommitNodeGraph are samed.

### Does this close any open issues?
close #1302 

### Current Behavior
Describe the current behaviour of this issue, if relevant.

### New Behavior
Describe the new behaviour updated in this issue, if relevant.

### Screenshots
Include any relevant screenshots here.

### Other Information
Any other information that is important to this PR.